Split synchronization mechanisms from util.{h,cpp}
Reference:
https://github.com/bitcoin/bitcoin/commit/7f3ccb59da31c7b1706ebfbb401910923221f076
