+++ src/makefile.linux-mingw
@@ -63,6 +63,7 @@ OBJS= \
     obj/rpcdump.o \
     obj/rpcnet.o \
     obj/script.o \
+    obj/sync.o \
     obj/util.o \
     obj/wallet.o \
     obj/walletdb.o \
