+++ src/makefile.mingw
@@ -116,6 +116,7 @@ OBJS= \
     obj/rpcdump.o \
     obj/rpcnet.o \
     obj/script.o \
+    obj/sync.o \
     obj/util.o \
     obj/wallet.o \
     obj/walletdb.o \
