+++ src/makefile.osx
@@ -97,6 +97,7 @@ OBJS= \
     obj/rpcdump.o \
     obj/rpcnet.o \
     obj/script.o \
+    obj/sync.o \
     obj/util.o \
     obj/wallet.o \
     obj/walletdb.o \
