+++ src/makefile.osx-mavericks
@@ -98,6 +98,7 @@ OBJS= \
     obj/rpcdump.o \
     obj/rpcnet.o \
     obj/script.o \
+    obj/sync.o \
     obj/util.o \
     obj/wallet.o \
     obj/walletdb.o \
